requirements.txt
pandas>=1.5.0
scikit-learn>=1.2.0
numpy>=1.24.0
python-dotenv>=1.0.0
dvc>=2.0.0
mlflow>=2.0.0
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0
matplotlib>=3.5.0
GitPython>=3.1.0